requirements.txt
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
eventkit==1.0.3
fonttools==4.53.1
ib-insync==0.9.86
idna==3.8
kiwisolver==1.4.5
matplotlib==3.9.2
nest-asyncio==1.6.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pyotp==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
robin-stocks==3.1.0
schedule==1.2.2
six==1.16.0
termcolor==2.4.0
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.2
